labs/07-multi-agent/demo-02-conditional-routing/pyproject.toml
[project]
name = "demo-02-conditional-routing"
version = "0.1.0"
description = "LangGraph conditional edges — route execution based on runtime state"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.0",
    "langchain>=1.0.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
]
